Revert "iap set-policy supports K8s Gateway resource"
This reverts commit cfe48ed9ca51ec4eedc90c0f3153bf94198dd1a6.

This was committed on the wrong branch
