+++ iap/resolve_test.go
@@ -1,16 +1,8 @@
 package iap
 
 import (
-	"context"
-	"os"
-	"path/filepath"
 	"testing"
 
-	"github.com/google/go-cmp/cmp"
-	"github.com/jlewi/monogo/k8s"
-	"k8s.io/client-go/util/homedir"
-
-	compute "cloud.google.com/go/compute/apiv1"
 	v1 "k8s.io/api/networking/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -58,42 +50,5 @@ func Test_GetGCPBackendFromIngress(t *testing.T) {
 				t.Errorf("Got %v; want %v", actual, c.Expected)
 			}
 		})
-	}
-}
-
-func Test_GatewayToBackend(t *testing.T) {
-	// This is an "integration" test that requires access to GCP
-	// It is useful for development. The actual values would have to be updated
-	// based on the project and service used for testing.
-	if os.Getenv("GITHUB_ACTIONS") != "" {
-		t.Skip("Skipping test in GitHub Actions")
-	}
-	bSvc, err := compute.NewBackendServicesRESTClient(context.Background())
-	defer bSvc.Close()
-
-	if err != nil {
-		t.Fatalf("Failed to create backend service client: %v", err)
-	}
-
-	k8sFlags := &k8s.K8SClientFlags{
-		Kubeconfig: filepath.Join(homedir.HomeDir(), ".kube", "config"),
-	}
-	k8sClient, err := k8sFlags.NewClient()
-
-	if err != nil {
-		t.Fatalf("Failed to create k8s client: %v", err)
-	}
-
-	negToBackend, err := GetGCPBackendFromService(k8sClient, bSvc, "chat-lewi", "gateway", "site-v1")
-	if err != nil {
-		t.Fatalf("Failed to get backend: %v", err)
-	}
-
-	expected := map[string]string{
-		"k8s1-ec8d20c1-gateway-site-v1-8080-c416a331": "gkegw1-u3ex-gateway-site-v1-8080-r546wxld07pe",
-	}
-
-	if d := cmp.Diff(expected, negToBackend); d != "" {
-		t.Errorf("Got unexpected diff: %v", d)
 	}
 }
