+++ oauthutil/firebase_flow.go
@@ -4,13 +4,8 @@ import (
 	"context"
 	"embed"
 	"encoding/json"
-	"fmt"
-	"io"
-	"net/http"
-	"os"
-	"time"
-
 	"firebase.google.com/go/auth"
+	"fmt"
 	"github.com/coreos/go-oidc/v3/oidc"
 	"github.com/go-logr/zapr"
 	"github.com/gorilla/mux"
@@ -20,6 +15,10 @@ import (
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 	"golang.org/x/oauth2"
+	"io"
+	"net/http"
+	"os"
+	"time"
 )
 
 // embed the assets
@@ -105,7 +104,9 @@ func (s *FirebaseFlowServer) NotFoundHandler(w http.ResponseWriter, r *http.Requ
 
 // waitForReady waits until the server is healthy.
 func (s *FirebaseFlowServer) waitForReady() error {
-	endTime := time.Now().Add(3 * time.Minute)
+	// DO NOT SUBMIT; the long timeout was for debugging
+	endTime := time.Now().Add(100 * time.Minute)
+	// endTime := time.Now().Add(3 * time.Minute)
 	for time.Now().Before(endTime) {
 
 		r, err := http.Get(s.Address() + "/healthz")
